+++ src/MainWindow/main_window.cpp
@@ -24,10 +24,10 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <QtWidgets>
 #include <fstream>
 
+#include "DesignRuns/runs_form.h"
 #include "Main/Foedag.h"
 #include "NewFile/new_file.h"
 #include "NewProject/new_project_dialog.h"
-#include "ProjNavigator/runs_form.h"
 #include "ProjNavigator/sources_form.h"
 #include "TextEditor/text_editor.h"
